requirements.txt
#
# Manually curated runtime dependencies for the RAG Assistant.
# These pins favor compatibility with Python 3.10 on Windows and ensure
# PyTorch/NumPy alignment for sentence-transformers.
#

# Application / UI
streamlit>=1.39.0,<2.0.0
python-dotenv>=1.0.1,<2.0.0
PyYAML>=6.0.2,<7.0.0

# LangChain stack
langchain-core>=0.3.78,<0.4.0
langchain-openai>=0.3.35,<0.4.0
langchain-groq>=0.3.8,<0.4.0
langchain-google-genai>=2.1.12,<3.0.0
langchain-text-splitters>=0.3.11,<0.4.0
openai>=1.104.2,<3.0.0
groq>=0.32.0,<1.0.0

# Retrieval / embeddings
chromadb>=0.5.23,<0.6.0
sentence-transformers>=3.2.1,<4.0.0
transformers>=4.45.2,<5.0.0
tiktoken>=0.7.0,<0.8.0

# Torch / numeric stack (pinned for compatibility)
# Note: NumPy 2.x requires recompiled modules. Using 1.x for compatibility.
torch>=2.2.2,<2.3.0
numpy>=1.26.0,<2.0.0
scikit-learn>=1.5.2,<1.6.0

# File loaders
python-docx>=1.1.2,<2.0.0
PyPDF2>=3.0.1,<4.0.0

# Misc utilities
huggingface-hub>=0.26.2,<0.27.0
requests>=2.32.3,<3.0.0
